# samples/jaxrs-spec/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Mail Baby API",
    "description" : "This is an API for accesssing the mail services.",
    "termsOfService" : "https://www.interserver.net/terms-of-service.html",
    "contact" : {
      "name" : "Joe Huss",
      "url" : "https://www.interserver.net/contact-information.html",
      "email" : "detain@interserver.net"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
...
  "servers" : [ {
    "url" : "http://mynew.interserver.net:8787/",
    "description" : "Temporary API Endpoint"
  }, {
    "url" : "https://api.interserver.net/",
    "description" : "Live API Endpoint"
  }, {
    "url" : "https://virtserver.swaggerhub.com/InterServer/MailBaby/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "security" : [ {
    "apiLoginAuth" : [ ],
    "apiPasswordAuth" : [ ]
  }, {
    "apiLoginAuth" : [ ],
    "apiKeyAuth" : [ ]
  } ],
  "paths" : {
...
    "/mail" : {
      "get" : {
        "summary" : "displays a list of mail service orders",
        "operationId" : "getMailOrders",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MailOrders"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MailOrders"
                }
              },
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "links" : {
              "getMailByIdLink" : {
                "operationId" : "getMailById",
                "parameters" : {
                  "id" : "$response.body#/0/id"
                },
                "description" : "The `id` value returned in the response can be used as the `id` parameter in `GET /mail/{id}`."
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/mail/order" : {
      "get" : {
        "summary" : "validatess order details before placing an order",
        "operationId" : "validateMailOrder",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "list of mail orders"
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "deprecated" : true
      },
      "post" : {
        "summary" : "places a mail order",
        "description" : "Adds an item to the system",
        "operationId" : "placeMailOrder",
        "parameters" : [ ],
        "requestBody" : {
          "description" : "Inventory item to add",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MailOrder"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "list of mail orders"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "an existing item already exists"
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "deprecated" : true
      }
    },
    "/mail/{id}" : {
      "get" : {
        "summary" : "Gets mail order information by id",
        "description" : "returns information about a mail order in the system with the given id.",
        "operationId" : "getMailById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "User ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MailOrder"
                }
              }
            }
          }
        }
      }
    },
    "/mail/{id}/send" : {
      "get" : {
        "summary" : "Sends an Email",
        "description" : "Sends An email through one of your mail orders.",
        "operationId" : "sendMailById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "User ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "subject",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "body",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/mail/{id}/log" : {
      "get" : {
        "summary" : "displays the mail log",
        "description" : "By passing in the appropriate options, you can search for\navailable inventory in the system\n",
        "operationId" : "viewMailLogById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "User ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "searchString",
          "in" : "query",
          "description" : "pass an optional search string for looking up inventory",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "maximum" : 50,
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MailLog"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        },
        "deprecated" : true
      }
    }
  },
  "components" : {
    "schemas" : {
      "Error" : {
        "required" : [ "code", "message" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "GenericResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "example" : "ok"
          },
          "status_text" : {
            "type" : "string",
            "example" : "The command completed successfully."
          }
        }
      },
      "MailOrders" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/MailOrder"
        }
      },
      "MailOrder" : {
        "required" : [ "id", "status", "username" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 1234
          },
          "status" : {
            "type" : "string",
            "example" : "active"
          },
          "username" : {
            "type" : "string",
            "example" : "mb1234"
          },
          "password" : {
            "type" : "string",
            "example" : "guest123"
          },
          "comment" : {
            "type" : "string",
            "example" : "main mail account"
          }
        }
      },
      "MailLog" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "description" : "Mail Order Details"
      }
    },
...
    "securitySchemes" : {
      "apiLoginAuth" : {
        "type" : "apiKey",
        "name" : "X-API-LOGIN",
        "in" : "header"
      },
      "apiPasswordAuth" : {
        "type" : "apiKey",
        "name" : "X-API-PASS",
        "in" : "header"
      },
      "apiKeyAuth" : {
        "type" : "apiKey",
        "name" : "X-API-KEY",
        "in" : "header"
      }
    },
...
  }
}